refactor: 🧱 Restructured the whole code base. All the styles moved into 'styles' folder. All the components are under one folder 'components'. All the pages are moved to same folder 'pages'.
