+++ src/components/colorbox.js
@@ -2,7 +2,7 @@ import { PureComponent } from 'react'
 import { Link } from 'react-router-dom';
 import { SlCopyButton, SlTooltip, SlIcon } from '@shoelace-style/shoelace/dist/react';
 import chroma from 'chroma-js';
-import './colorbox.css'
+import '../styles/colorbox.css'
 
 class ColorBox extends PureComponent {
   render () {
