+++ src/styles/miniPalette.css
