+++ src/styles/palette.css
